+++ src/LlmTornado.Agents/DataModels/GuardRail.cs
