+++ src/LlmTornado.Agents/DataModels/ModelOutputFormat.cs
