+++ src/LlmTornado.Agents/DataModels/ModelResponse.cs
@@ -32,21 +32,12 @@
 
         public ModelResponse() { }
 
-        [Obsolete]
-        public ModelResponse(string model, List<ModelItem> outputItems = null, ModelOutputFormat outputFormat = null, List<ModelItem> messages = null)
-        {
-            Model = model;
-            OutputItems = outputItems;
-            OutputFormat = outputFormat;
-            Messages = messages ?? new List<ModelItem>();
-        }
-
-        public ModelResponse(List<ModelItem> outputItems = null, ModelOutputFormat outputFormat = null, List<ModelItem> messages = null, string? id = null)
+        public ModelResponse(List<ModelItem> outputItems = null, ModelOutputFormat outputFormat = null, List<ModelItem> messages = null, string? responseId = null)
         {
             OutputItems = outputItems;
             OutputFormat = outputFormat;
             Messages = messages ?? new List<ModelItem>();
-            Id = id;
+            Id = responseId;
         }
     }
 }
