+++ src/LlmTornado.Agents/DataModels/ModelResponseContent.cs
