+++ src/LlmTornado.Agents/DataModels/ModelResponseItems.cs
