+++ src/LlmTornado.Agents/DataModels/RunResult.cs
