+++ src/LlmTornado.Agents/Utility/CommonUtility.cs
