+++ src/LlmTornado.Agents/Utility/ComputerToolUtility.cs
