+++ src/LlmTornado.Agents/Utility/json_util.cs
